AGNT-1218 Release iOS SDK for PhoneGap
