+++ plugin.xml
@@ -43,7 +43,8 @@
         <header-file src="platforms/ios/CrittercismSDK/Crittercism.h" />
         <header-file src="platforms/ios/CrittercismSDK/CrittercismConfig.h" />
         <header-file src="platforms/ios/CrittercismSDK/CrittercismDelegate.h" />
-        <source-file src="platforms/ios/CrittercismSDK/libCrittercism_v5_3_1.a" framework="true" />
+        <header-file src="platforms/ios/CrittercismSDK/CrittercismLoggingLevel.h" />
+        <source-file src="platforms/ios/CrittercismSDK/libCrittercism_v5_4_1.a" framework="true" />
         <source-file src="platforms/ios/CrittercismSDK/dsym_upload.sh" />
 
         <config-file target="*-Info.plist" parent="CritterAppID">
